Fixes #1261: amount string is now converted to a double to avoid overflow

+++ mifospay/src/main/java/org/mifos/mobilewallet/mifospay/standinginstruction/ui/NewSIActivity.kt
@@ -78,7 +78,7 @@ class NewSIActivity : BaseActivity(), StandingInstructionContract.NewSIView {
         if (et_si_amount.text.toString() == "") {
             showToast(getString(R.string.enter_amount))
             return
-        } else if (et_si_amount.text.toString().toInt() <= 0) {
+        } else if (et_si_amount.text.toString().toDouble() <= 0) {
             showToast(getString(R.string.enter_valid_amount))
             return
         }
